Add files via upload
Write an application that inputs one number consisting of five digits 
from the user, separates the number into its individual digits and prints the 
digits separated from one another by three spaces each. For example, if the
user types in the number 42339, the program should print “4 2 3 3 9.” Assume that the user enters
the correct number of digits.